feat: moved gql client and  qeury
